Rebase and handle feedback
