+++ pkg/storage/endpoint/tso.go
@@ -50,7 +50,6 @@ func (se *StorageEndpoint) LoadTimestamp(prefix string) (time.Time, error) {
 	maxTSWindow := typeutil.ZeroTime
 	for i, key := range keys {
 		key := strings.TrimSpace(key)
-		log.Info("LoadTimestamp", zap.String("ts-window-key", key))
 		if !strings.HasSuffix(key, timestampKey) {
 			continue
 		}
@@ -63,7 +62,6 @@ func (se *StorageEndpoint) LoadTimestamp(prefix string) (time.Time, error) {
 			maxTSWindow = tsWindow
 		}
 	}
-	log.Info("LoadTimestamp", zap.Time("max-ts-window", maxTSWindow))
 	return maxTSWindow, nil
 }
